Update Analysis, RefCount, and Exceptions Tests
Additionally, refactor Genops tests by moving shared functions into
testutil.py.
