+++ test-data/refcount.test
@@ -4,6 +4,8 @@
 def f() -> int:
     return 1
 [out]
+def f():
+    r0 :: int
 L0:
     r0 = 1
     return r0
@@ -13,6 +15,8 @@ def f() -> int:
     x = 1
     return x
 [out]
+def f():
+    r0, x :: int
 L0:
     r0 = 1
     x = r0
@@ -25,6 +29,8 @@ def f() -> int:
     x = y
     return x
 [out]
+def f():
+    r0, x, y :: int
 L0:
     r0 = 1
     x = r0
@@ -39,6 +45,8 @@ def f() -> int:
     z = x
     return y + z
 [out]
+def f():
+    r0, x, y, z, r1 :: int
 L0:
     r0 = 1
     x = r0
@@ -58,6 +66,9 @@ def f() -> int:
         return x
     return y
 [out]
+def f():
+    r0, x, r1, y, r2 :: int
+    r3 :: bool
 L0:
     r0 = 1
     x = r0
@@ -84,6 +95,8 @@ def f(a: int, b: int) -> int:
     y = x + a
     return y
 [out]
+def f(a, b):
+    a, b, r0, r1, x, r2, y :: int
 L0:
     r0 = 1
     r1 = a + r0 :: int
@@ -101,6 +114,8 @@ def f(a: int) -> int:
     x = 1
     return x + y
 [out]
+def f(a):
+    a, x, y, r0, r1 :: int
 L0:
     inc_ref a :: int
     x = a
@@ -120,6 +135,8 @@ def f(a: int) -> int:
     y = a
     return y
 [out]
+def f(a):
+    a, r0, y :: int
 L0:
     r0 = 1
     a = r0
@@ -133,6 +150,8 @@ def f(a: int) -> int:
     a = 3
     return a
 [out]
+def f(a):
+    a, r0, r1, r2 :: int
 L0:
     r0 = 1
     a = r0
@@ -151,6 +170,8 @@ def f(a: int) -> int:
     y = x
     return a
 [out]
+def f(a):
+    a, r0, x, y :: int
 L0:
     r0 = 1
     x = r0
@@ -164,6 +185,8 @@ L0:
 def f(a: int) -> int:
     return a
 [out]
+def f(a):
+    a :: int
 L0:
     inc_ref a :: int
     return a
@@ -177,6 +200,10 @@ def f(a: int) -> int:
     y = a + 1
     return y
 [out]
+def f(a):
+    a :: int
+    r0 :: bool
+    r1, r2, x, r3, r4, y :: int
 L0:
     r0 = a == a :: int
     if r0 goto L1 else goto L2 :: bool
@@ -209,6 +236,10 @@ def f(a: int) -> int:
     y = a + 1
     return y
 [out]
+def f(a):
+    a :: int
+    r0 :: bool
+    r1, x, r2, r3, r4, y :: int
 L0:
     r0 = a == a :: int
     if r0 goto L1 else goto L2 :: bool
@@ -237,6 +268,10 @@ def f(a: int) -> int:
         a = 1
     return a
 [out]
+def f(a):
+    a :: int
+    r0 :: bool
+    r1 :: int
 L0:
     r0 = a == a :: int
     if r0 goto L1 else goto L3 :: bool
@@ -257,6 +292,8 @@ def f(a: int) -> int:
     return x + a
 -- This is correct but bad code
 [out]
+def f(a):
+    a, r0, x, r1 :: int
 L0:
     inc_ref a :: int
     a = a
@@ -277,6 +314,8 @@ def f(a: int) -> int:
     x = x + 1
     return a + x
 [out]
+def f(a):
+    a, r0, r1, r2, x, r3, r4, r5 :: int
 L0:
     r0 = 1
     r1 = a + r0 :: int
@@ -299,6 +338,9 @@ def f() -> None:
     x = 1
     x = x + 1
 [out]
+def f():
+    r0, x, r1, r2 :: int
+    r3 :: None
 L0:
     r0 = 1
     x = r0
@@ -317,6 +359,9 @@ def f() -> None:
     y = 1
     x = y + 1
 [out]
+def f():
+    r0, y, r1, r2, x :: int
+    r3 :: None
 L0:
     r0 = 1
     y = r0
@@ -337,6 +382,8 @@ def f(a: int) -> int:
     x = x + x
     return x
 [out]
+def f(a):
+    a, r0, x, r1 :: int
 L0:
     r0 = a + a :: int
     a = r0
@@ -352,6 +399,8 @@ def f(a: int) -> int:
     y = x + x
     return y
 [out]
+def f(a):
+    a, r0, x, r1, y :: int
 L0:
     r0 = a + a :: int
     x = r0
@@ -366,6 +415,9 @@ def f(a: int) -> None:
     y = 1
     z = y + y
 [out]
+def f(a):
+    a, r0, x, r1, y, r2, z :: int
+    r3 :: None
 L0:
     r0 = a + a :: int
     x = r0
@@ -386,6 +438,9 @@ def f(a: int) -> None:
     x = 1
     x = x + x
 [out]
+def f(a):
+    a, r0, r1, x, r2 :: int
+    r3 :: None
 L0:
     r0 = a + a :: int
     a = r0
@@ -410,6 +465,10 @@ def f() -> int:
     a = 1
     return x + y - a
 [out]
+def f():
+    r0, x, r1, y, r2, z :: int
+    r3 :: bool
+    r4, a, r5, r6 :: int
 L0:
     r0 = 1
     x = r0
@@ -448,6 +507,10 @@ def f(a: int) -> int:
         i = i + 1
     return sum
 [out]
+def f(a):
+    a, r0, sum, r1, i :: int
+    r2 :: bool
+    r3, r4, r5 :: int
 L0:
     r0 = 0
     sum = r0
@@ -476,6 +539,8 @@ L4:
 def f(a: int) -> int:
     return f(a + 1)
 [out]
+def f(a):
+    a, r0, r1, r2 :: int
 L0:
     r0 = 1
     r1 = a + r0 :: int
@@ -492,6 +557,11 @@ def f() -> int:
     a = [0, 1]
     return 0
 [out]
+def f():
+    r0, r1 :: int
+    r2, r3 :: object
+    r4, a :: list
+    r5 :: int
 L0:
     r0 = 0
     r1 = 1
@@ -510,6 +580,10 @@ from typing import List
 def f(x: int) -> List[int]:
     return [x]
 [out]
+def f(x):
+    x :: int
+    r0 :: object
+    r1 :: list
 L0:
     inc_ref x :: int
     r0 = box(int, x)
@@ -522,6 +596,14 @@ from typing import List
 def f(a: List[int], b: List[int]) -> None:
     a[0] = b[0]
 [out]
+def f(a, b):
+    a, b :: list
+    r0 :: int
+    r1 :: object
+    r2, r3 :: int
+    r4 :: object
+    r5 :: bool
+    r6 :: None
 L0:
     r0 = 0
     r1 = b[r0] :: list
@@ -542,6 +624,10 @@ from typing import Tuple
 def f(x: Tuple[Tuple[int, bool], bool]) -> int:
     return x[0][0]
 [out]
+def f(x):
+    x :: tuple[tuple[int, bool], bool]
+    r0 :: tuple[int, bool]
+    r1 :: int
 L0:
     r0 = x[0]
     r1 = r0[0]
@@ -555,6 +641,10 @@ def f() -> None:
     c = C()
     c.x = C()
 [out]
+def f():
+    r0, c, r1 :: C
+    r2 :: bool
+    r3 :: None
 L0:
     r0 = C()
     c = r0
@@ -573,6 +663,13 @@ def f() -> None:
     a = [C()]
     d = a[0]
 [out]
+def f():
+    r0 :: C
+    r1, a :: list
+    r2 :: int
+    r3 :: object
+    r4, d :: C
+    r5 :: None
 L0:
     r0 = C()
     r1 = [r0]
@@ -595,6 +692,9 @@ def f(x: bool) -> int:
         return 1
     return 2
 [out]
+def f(x):
+    x :: bool
+    r0, r1 :: int
 L0:
     if x goto L1 else goto L2 :: bool
 L1:
@@ -608,6 +708,8 @@ L2:
 def f() -> str:
     return "some string"
 [out]
+def f():
+    r0 :: str
 L0:
     r0 = unicode_0 :: static  ('some string')
     inc_ref r0
@@ -617,6 +719,19 @@ L0:
 def g(x: str) -> int:
     return int(x, base=2)
 [out]
+def g(x):
+    x :: str
+    r0 :: int
+    r1 :: object
+    r2 :: str
+    r3 :: object
+    r4 :: str
+    r5 :: tuple[str]
+    r6 :: dict
+    r7 :: object
+    r8 :: bool
+    r9, r10 :: object
+    r11 :: int
 L0:
     r0 = 2
     r1 = builtins.module :: static
@@ -642,6 +757,12 @@ from typing import List
 def f(a: List[int], x: int) -> None:
     a.append(x)
 [out]
+def f(a, x):
+    a :: list
+    x :: int
+    r0 :: object
+    r1 :: bool
+    r2, r3 :: None
 L0:
     inc_ref x :: int
     r0 = box(int, x)
@@ -659,6 +780,14 @@ def f(d: Dict[int, int]) -> None:
     for key in d:
         d[key]
 [out]
+def f(d):
+    d :: dict
+    r0, r1 :: object
+    key, r2 :: int
+    r3, r4 :: object
+    r5 :: int
+    r6 :: bool
+    r7 :: None
 L0:
     r0 = iter d :: object
 L1:
